UniversalToJsonFormatter and AbstractSchemaBasedRecordFormatter flatten
